// src/app/shared/model/log.model.ts
export interface Log {
     id?: string; // identificador único
     cod?: string; // máximo 15 caractere
     display?: string; // máximo 25 caracteres
     tipo?: string; // máximo 250 caracteres
     tipo2?: string; // máximo 60 caracteres
     updatedAt?: any;
     createdAt?: any;
     4?: string;
     5?: string;
     6?: string;
     7?: string;
     8?: string;
     9?: string;
     10?: string;
     11?: string;
     12?: string;
     13?: string;
     14?: string
  }